requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
flake8
pytest
sqlalchemy
httpx
prometheus-client
prometheus-fastapi-instrumentator